refactor: Correção de alguns codigos para funcionamento de carrinho

+++ app/src/main/java/br/ufv/tp1_poo/model/Vegetariano.java
@@ -34,8 +34,8 @@ public class Vegetariano extends Produto {
         }
     }
     @Override
-    public int calculaPreco() {
-        return ((int)this.getPreco() + this.getAdicionalPorTamanho()) * this.getQuantidade();
+    public Float calculaPreco() {
+        return ((float)this.getPreco() + this.getAdicionalPorTamanho()) * this.getQuantidade();
     }
 
     private int getAdicionalPorTamanho() {
